fix(controllers): share utilization backoff decision
